✏️ Move to Hero Component
